Firebase Initialized
PawDates is now connected to my Firebase account from where we will host
the site. All of your stuff goes into the pawdates folder so that we
don't have to worry (right now) about file paths. Front page goes on
index.html, and if we want a custom 404 page (it'd be really cool), that
goes in 404.html.