Add init script and db structure example
I added the initialization code to index.html. Please don't mess with
it, as it is the only thing giving functionality to our app. Check out
the example database and see if I need to add anything. The structure is
pretty much exactly like dictionaries in Python with keys and values,
but you can have another dictionary as a value. Wherever a value and a
key are the same (including the dogs array in the user entry), that's
the two-way redundancy I was talking about last Thursday, if you were
paying attention (I wasn't).